requirements.txt
python-dotenv~=1.0.0
revChatGPT~=2.3.21
telebot~=0.0.5
openai~=0.27.0
pytesseract~=0.3.10
Pillow~=9.4.0